Fixing null exception bug

+++ project2/restaurant-parent/client/src/main/java/com/cpre550/project2/restaurant/client/Client.java
@@ -43,7 +43,7 @@ public class Client{
 			System.out.print("> ");
 			ui.parse_cmd(scanner.nextLine());
 		}
-		daemon.kill();
+		if(daemon != null) daemon.kill();
 		scanner.close();
     }
